Restrict access to role command

+++ cmds/role/createMessage.go
@@ -19,12 +19,14 @@ var CreateMessageRoleCommand = &cmdlr2.Command{
 			Fields:      []*disgord.EmbedField{},
 		}
 
-		m, _ := ctx.Client.Channel(ctx.Event.Message.ChannelID).CreateMessage(&disgord.CreateMessageParams{Embed: embed})
+		if ctx.Event.Message.Author.ID == 133314498214756352 || ctx.Event.Message.Author.ID == 271787171889807360 {
+			m, _ := ctx.Client.Channel(ctx.Event.Message.ChannelID).CreateMessage(&disgord.CreateMessageParams{Embed: embed})
 
-		err := db.CreateRoleMessage(int64(m.ID), int64(ctx.Event.Message.ChannelID), int64(ctx.Event.Message.GuildID))
-		if err != nil {
-			_ = ctx.Client.Channel(ctx.Event.Message.ChannelID).DeleteMessages(&disgord.DeleteMessagesParams{Messages: []disgord.Snowflake{m.ID}})
-			ctx.ResponseText(fmt.Sprintf("%v", err))
+			err := db.CreateRoleMessage(int64(m.ID), int64(ctx.Event.Message.ChannelID), int64(ctx.Event.Message.GuildID))
+			if err != nil {
+				_ = ctx.Client.Channel(ctx.Event.Message.ChannelID).DeleteMessages(&disgord.DeleteMessagesParams{Messages: []disgord.Snowflake{m.ID}})
+				ctx.ResponseText(fmt.Sprintf("%v", err))
+			}
 		}
 	},
 }
